Add socket to paymentController
Fill in the blanks @leesuby

+++ common/src/main/java/com/seasidechachacha/common/payment/NewUserRequest.java
@@ -22,7 +22,7 @@ public class NewUserRequest implements Serializable {
         this.userId = userId;
     }
 
-    public int getDeposit() {
+    public double getDeposit() {
         return deposit;
     }
